requirements.txt
streamlit
torch
torchvision
Pillow
rapidfuzz
streamlit-cropper